Constrain custom section annotation placement
The overview previously did not constrain the location of custom section annotations, implying that they should be allowed anywhere annotations in general are allowed, i.e. anywhere whitespace is allowed.

This flexibility is not necessary since the location of custom section annotations is not meaningful, and it also complicates parser implementations. To allow for simpler implementations that only check for custom section annotations in a single location, constrain the annotations to appear only at the top level of the module.